+++ alphastats/dataset/factory.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
-from dataset.dataset_harmonizer import DataHarmonizer
+from dataset.harmonizer import DataHarmonizer
 
 from alphastats.keys import Cols
