+++ alphastats/gui/pages/03_Preprocessing.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from dataset.DataSet_Preprocess import PreprocessingStateKeys
+from dataset.preprocessing import PreprocessingStateKeys
 
 from alphastats.gui.utils.preprocessing_helper import (
     configure_preprocessing,
