+++ alphastats/gui/utils/overview_helper.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import streamlit as st
-from dataset.DataSet import DataSet
+from dataset.dataset import DataSet
 
 from alphastats.gui.utils.ui_helper import StateKeys, show_button_download_df
