+++ tests/llm/test_llm_functions.py
@@ -8,7 +8,7 @@ import inspect
 from typing import Callable, Dict
 
 import pandas as pd
-from dataset.DataSet import DataSet
+from dataset.dataset import DataSet
 
 from alphastats.llm.llm_functions import (
     GENERAL_FUNCTION_MAPPING,
