+++ tests/test_DataSet.py
@@ -9,9 +9,9 @@ from unittest.mock import patch
 import numpy as np
 import pandas as pd
 import plotly
-from dataset.DataSet import DataSet
-from dataset.dataset_factory import DataSetFactory
-from dataset.DataSet_Preprocess import PreprocessingStateKeys
+from dataset.dataset import DataSet
+from dataset.factory import DataSetFactory
+from dataset.preprocessing import PreprocessingStateKeys
 
 from alphastats.loader.alphapept_loader import AlphaPeptLoader
 from alphastats.loader.diann_loader import DIANNLoader
