+++ tests/test_DataSet_Pathway.py
@@ -2,8 +2,8 @@ import unittest
 
 import numpy as np
 import pandas as pd
-from dataset.DataSet import DataSet
-from dataset.DataSet_Pathway import Enrichment
+from dataset.dataset import DataSet
+from dataset.pathway import Enrichment
 
 from alphastats.loader.maxquant_loader import MaxQuantLoader
